Move common regular expressions to a separate file to avoid dependency erros
